Allow the parallelism of stream loading to be defined in the configuration
If running in a container on a large host Python reports the number of CPUs as the number of host CPUS, see
https://bugs.python.org/issue36054
this means that there's a large number of threads spawned to process streams for a very small number of VCPUs which have been allocated.

This adds a 'parallelism' flag indicating the number of streams which will be processed in parallel to work around this issue.