Allow S3 COPY role arn to be provided, remove required AWS credentials
Allow credentials to be taken from the runtime environment by default. These are fetched by boto3.

This allows the user to deploy the target or pipelinewise with the correct environment and not be forced to mint keys or tokens from a role arn, e.g. running on EC2/ECS.

Additionally allow a role ARN to be provided via aws_redshift_copy_role_arn which will allow the user to use environment credentials and delegate the COPY command to a role
which redshift can assume.