Add 'slices' configuration option
This allows the user to configure the number of chunks that files will
be loaded in. This should improve parallel loading.

It may be sensible to also add a minimum chunk size. The recommended
minimum/maximum sizes are 1MB/256MB compressed, however I'm not sure how
to best implement this automatically in a quick and sensible way.